+++ examples/simple-taro-fast-framework/src/news/pages/Home2/Main/index.jsx
@@ -74,7 +74,7 @@ export default class Index extends BasePageWrapper {
     });
   };
 
-  afterLoadSuccess = ({
+  doOtherAfterLoadSuccess = ({
     // eslint-disable-next-line no-unused-vars
     metaData = null,
     // eslint-disable-next-line no-unused-vars