+++ examples/simple-template/src/modelBuilders/general/flowCase.js
@@ -1,12 +1,14 @@
 import {
   getTacitlyState,
   pretreatmentRemotePageListData,
+  pretreatmentRemoteSingleData,
   reducerCollection,
   reducerDefaultParameters,
   reducerNameCollection,
 } from 'easy-soft-utility';
 
 import {
+  getData,
   pageListLatestApproveData,
   pageListWaitApproveData,
 } from '../../services/flowCase';
@@ -14,6 +16,7 @@ import {
 export const flowCaseTypeCollection = {
   pageListWaitApprove: 'flowCase/pageListWaitApprove',
   pageListLatestApprove: 'flowCase/pageListLatestApprove',
+  get: 'flowCase/get',
 };
 
 export function buildModel() {
@@ -63,6 +66,32 @@ export function buildModel() {
         const response = yield call(pageListLatestApproveData, payload);
 
         const dataAdjust = pretreatmentRemotePageListData({
+          source: response,
+          successCallback: pretreatmentSuccessCallback || null,
+          failCallback: pretreatmentFailCallback || null,
+        });
+
+        yield put({
+          type: reducerNameCollection.reducerRemoteData,
+          payload: dataAdjust,
+          alias,
+          ...reducerDefaultParameters,
+        });
+
+        return dataAdjust;
+      },
+      *get(
+        {
+          payload,
+          alias,
+          pretreatmentSuccessCallback,
+          pretreatmentFailCallback,
+        },
+        { call, put },
+      ) {
+        const response = yield call(getData, payload);
+
+        const dataAdjust = pretreatmentRemoteSingleData({
           source: response,
           successCallback: pretreatmentSuccessCallback || null,
           failCallback: pretreatmentFailCallback || null,