+++ examples/simple-template/CHANGELOG.md
@@ -1,5 +1,17 @@
 # taro-fast-framework-template
 
+## 1.0.8
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - taro-fast-component@3.0.14
+  - taro-fast-framework@3.1.8
+  - taro-fast-common@3.1.8
+  - taro-fast-component-extra@3.0.14
+  - taro-fast-component-prism@3.0.15
+
 ## 1.0.7
 
 ### Patch Changes